+++ CHANGELOG.md
@@ -13,7 +13,7 @@
 
 ### Improved
 
-- Upgrade `python` to 3.8.2.
+- Upgrade `python` to 3.8.5.
 - All requirements upgraded to latest version.
 
 ### Added